pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name="heart-disease"
authors=[
    {name="Nikhil Kudupudi", email="nikhilkudupudi@gmail.com"}
]
dynamic=["version"]
dependencies=[
    "mysql-connector-python",
    "python-dotenv",
    "datetime",
    "flask",
   
]
requires-python = ">= 3.8"
readme = "README.md"
